ref: Move sentry options into its separate file
In preparation for more fixes
