+++ src/backends/sentry_backend_breakpad.cpp
@@ -4,7 +4,9 @@ extern "C" {
 #include "sentry_alloc.h"
 #include "sentry_backend.h"
 #include "sentry_core.h"
+#include "sentry_database.h"
 #include "sentry_envelope.h"
+#include "sentry_options.h"
 #include "sentry_path.h"
 #include "sentry_sync.h"
 #include "sentry_transport.h"
