+++ src/sentry_core.h
@@ -3,10 +3,6 @@
 
 #include "sentry_boot.h"
 
-#include "sentry_database.h"
-#include "sentry_path.h"
-#include "sentry_utils.h"
-
 #define SENTRY_BREADCRUMBS_MAX 100
 
 #ifdef SENTRY_PLATFORM_ANDROID
@@ -45,56 +41,6 @@
 #else
 #    define UNUSED(x) UNUSED_##x
 #endif
-
-struct sentry_backend_s;
-
-/**
- * This is a linked list of all the attachments registered via
- * `sentry_options_add_attachment`.
- */
-typedef struct sentry_attachment_s sentry_attachment_t;
-struct sentry_attachment_s {
-    char *name;
-    sentry_path_t *path;
-    sentry_attachment_t *next;
-};
-
-/**
- * This is the main options struct, which is being accessed throughout all of
- * the sentry internals.
- */
-struct sentry_options_s {
-    char *raw_dsn;
-    sentry_dsn_t dsn;
-    double sample_rate;
-    char *release;
-    char *environment;
-    char *dist;
-    char *http_proxy;
-    char *ca_certs;
-    sentry_path_t *database_path;
-    sentry_path_t *handler_path;
-    bool debug;
-    bool require_user_consent;
-    bool system_crash_reporter_enabled;
-
-    sentry_attachment_t *attachments;
-    sentry_run_t *run;
-
-    sentry_transport_t *transport;
-    sentry_event_function_t before_send_func;
-    void *before_send_data;
-
-    /* everything from here on down are options which are stored here but
-       not exposed through the options API */
-    struct sentry_backend_s *backend;
-    sentry_user_consent_t user_consent;
-};
-
-/**
- * This will free a previously allocated attachment.
- */
-void sentry__attachment_free(sentry_attachment_t *attachment);
 
 /**
  * This function will check the user consent, and return `true` if uploads
